+++ packages/react-mui-store-file-viewer/src/stories/StoreFileViewer.stories.tsx
@@ -278,6 +278,9 @@ StoreFileViewerDemo.args = {
     alerts: {
       displayAlertOnError: true,
     },
+    settings: {
+      maxAccessCount: 15,
+    },
   },
   defaultShareLinkTitle: "Pepe Silvia has securely shared a file with you!",
 };